backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
python-multipart
pillow
pytest
httpx
coverage
mypy==1.9.0
black>=24.0.0
